refactor: reorganize sidebar components and update LibrarySongItem
- Moved LibrarySongItem to a new directory structure under `library/`.
- Deleted the old LibrarySongItem component and created a new one in the new location.
- Updated Sidebar to import Library from the new path.
- Modified SidebarFooter to correct the spelling of "Uriyyá".
- Removed obsolete tests for Library and LibrarySongItem, and added new tests for the reorganized components.
- Updated the Library component to handle library path management and song filtering.
