+++ src/components/sidebar/SidebarFooter.tsx
@@ -1,7 +1,7 @@
 export function SidebarFooter() {
     return (
         <div className="sidebar-footer">
-            <div>Uriyya Music Player v0.4.0</div>
+            <div>Uriyyá Music Player v0.4.0</div>
             <div>Hecho con ❤️ por <a href="https://github.com/bhbarquero-dev">bhbarquero-dev</a></div>
         </div>
     );
